+++ openharmony/arkui_ace_engine/frameworks/bridge/cj_frontend/interfaces/cj_ffi/cj_shape_ffi.cpp
@@ -172,7 +172,7 @@ void FfiOHOSAceFrameworkShapeSetMesh(VectorFloat64Ptr vecValue, uint32_t column,
         LOGE("mesh array is empty");
         return;
     }
-    auto meshValue = *reinterpret_cast<std::vector<double>*>(vecValue);
+    auto meshValue = *reinterpret_cast<std::vector<float>*>(vecValue);
     ShapeModel::GetInstance()->SetBitmapMesh(meshValue, static_cast<int32_t>(column), static_cast<int32_t>(row));
 }
