+++ openharmony/arkui_ace_engine/frameworks/core/interfaces/native/node/shape_modifier.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Huawei Device Co., Ltd.
+ * Copyright (c) 2023-2024 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -44,7 +44,7 @@ void SetShapeMesh(ArkUINodeHandle node, const ArkUI_Float32* mesh, ArkUI_Uint32 
 {
     auto* frameNode = reinterpret_cast<FrameNode*>(node);
     CHECK_NULL_VOID(frameNode);
-    std::vector<double> meshValues(mesh, mesh + arrayItemCount);
+    std::vector<float> meshValues(mesh, mesh + arrayItemCount);
     ShapeModelNG::SetBitmapMesh(frameNode, meshValues, column, row);
 }
 
@@ -52,7 +52,7 @@ void ResetShapeMesh(ArkUINodeHandle node)
 {
     auto* frameNode = reinterpret_cast<FrameNode*>(node);
     CHECK_NULL_VOID(frameNode);
-    std::vector<double> meshValues;
+    std::vector<float> meshValues;
     int32_t column = 0;
     int32_t row = 0;
     ShapeModelNG::SetBitmapMesh(frameNode, meshValues, column, row);
